Add modified segments tracking - `get` never marks segment `modified` - `set` always marks segment `modified` - `delete` marks segment modified only if key was actually deleted (no mark for missing keys)
For `delete` it was easy to track only actual changes because it's just about checking boolean value.
For `set` it's much harder because we need to make `get` internally,
compare existing data with data we will replace it with and only then we can make a decision about
marking segment is modified. I think that potential extra save is much cheaper operation than
computational cost of making additional get and potential comparison of very large byte arrays
for each `set` call.

Misc:
- removed redundant `get` call from `set` method