Add dummy workload to test storage
Rationale: needed something to emulate storage usage and the only
existing controller is broken and always returns 4XX code and using it
would also require additional setup to make this workload `external`

For now it's a single task (but designed to be extensible).

Misc:
- fix log sentence scope inside NamespaceBackgroundSave (moved out of loop).
- fix creating directories for namespaces.
- fix deadlock when thread requesting existing namespace with
getOrCreateNamespace was blocking other threads from using method.